:art: Improving structure / format of Cat feature
Improving structure / format of Cat feature
